Refactor functions to create cobra commands into separate module
* Makes it easier to reuse logic across multiple binaries.
